refactor(src/main): 오버플로우 수정
기존 int 타입으로는 상금이 21억이 넘어가면 오버플로우가 발생하는 문제로
인해 상금 관련 데이터 타입을 long으로 수정

BREAKING CHANGE: 상금 관련 데이터 타입을 int에서 long으로 수정
LotteryGame:
as-is:
public class LotteryGame {
    ...

    public double makeRankingPercent(final Map<Rank, Integer> rankResult) {
        int earningAmount = 0;
        ...
    }

    ...
}

to-be:
public double makeRankingPercent(final Map<Rank, Integer> rankResult) {
    long earningAmount = 0;
    ...
}

PurchaseInformation:
as-is:
public class PurchaseInformation {
    ...

    public double calculateEarningRate(final int earningAmount) {
        ...
    }
}

to-be:
public class PurchaseInformation {
    ...

    public double calculateEarningRate(final long earningAmount) {
        ...
    }
}

PurchaseAmount;
as-is:
public class PurchaseAmount {
    ...

    public double calculateEarningRate(final int earningAmount) {
        ...
    }
}

to-be:
public class PurchaseAmount {
    ...

    public double calculateEarningRate(final long earningAmount) {
        ...
    }
}

Rank:
as-is:
public enum Rank {
    ...

    private final int prize;

    ...

    public int getPrize() {
        ...
    }

    ...
}

to-be:
public enum Rank {
    ...

    private final long prize;

    ...

    public long getPrize() {
        ...
    }

    ...
}

RankDto:
as-is:
public class RankDto {
    ...
    private final int prize;

    private RankDto(final int correctedBalls, final int prize, final String rankName) {
        ...
        this.prize = prize;
        ...
    }

    ...

    public int getPrize() {
        return prize;
    }

    ...
}

to-be:
public class RankDto {
     ...
    private final long prize;

    private RankDto(final int correctedBalls, final long prize, final String rankName) {
        ...
        this.prize = prize;
        ...
    }

    ...

    public long getPrize() {
        return prize;
    }

    ...

}

+++ src/main/java/controller/dto/RankDto.java
@@ -5,10 +5,10 @@ import domain.Rank;
 public class RankDto {
 
 	private final int correctedBalls;
-	private final int prize;
+	private final long prize;
 	private final String rankName;
 
-	private RankDto(final int correctedBalls, final int prize, final String rankName) {
+	private RankDto(final int correctedBalls, final long prize, final String rankName) {
 		this.correctedBalls = correctedBalls;
 		this.prize = prize;
 		this.rankName = rankName;
@@ -22,7 +22,7 @@ public class RankDto {
 		return correctedBalls;
 	}
 
-	public int getPrize() {
+	public long getPrize() {
 		return prize;
 	}
